requirements.txt
altair==5.5.0
annotated-types==0.7.0
anthropic==0.52.0
anyio==4.9.0
argcomplete==3.6.2
attrs==25.3.0
blinker==1.9.0
boto3==1.38.23
botocore==1.38.23
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cohere==5.15.0
colorama==0.4.6
Deprecated==1.2.18
diskcache==5.6.3
distro==1.9.0
eval_type_backport==0.2.2
fasta2a==0.2.9
fastavro==1.11.1
filelock==3.18.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.2
google-genai==1.16.1
graphiti-core==0.11.6
griffe==1.7.3
groq==0.25.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.1
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
logfire-api==3.16.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.9.1
mdurl==0.1.2
mistralai==1.7.1
narwhals==1.40.0
neo4j==5.28.1
numpy==2.2.6
openai==1.82.0
opentelemetry-api==1.33.1
packaging==24.2
pandas==2.2.3
pillow==11.2.1
prompt_toolkit==3.0.51
protobuf==6.31.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic-ai==0.2.9
pydantic-ai-slim==0.2.9
pydantic-evals==0.2.9
pydantic-graph==0.2.9
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==2.3.5
starlette==0.46.2
streamlit==1.45.1
tenacity==9.1.2
tokenizers==0.21.1
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
types-requests==2.32.0.20250515
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
zipp==3.21.0